pyproject.toml
[tool.poetry]
name = "educational-modules"
version = "0.1.0"
description = ""
authors = ["Dontgoingforyou <mishael000@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
flake8 = "^7.1.1"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.4.0"
drf-yasg = "^1.21.8"
django = "^5.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
